Move text utils to strings utils
